Add 'boundaryNode'.
Pretty straightforward. The function checks whether the given node is a
boundary not and returns 'STM False' if not and 'STM True' if it is.
The return value has to be wrapped in the STM monad since the node's
neighbours have to be unwrapped to test wheter the node is on the
boundary or not.