Add 'Config' type.
The type acts as an everything bucket for the parameters which need to
be passed down to the various functions executed to do a threaded run of
a GSOM phase. It should probably be refactored to a State monad and
split into units which belong together semantically but it works for
now, although I know it's ugly.