Move creation of some 'Config' entries to 'phase'.
This fixes two of bugs. 
First, 'step' is initialized in the right place, when a phase is
started. Initializing it in 'pass' made it get re-initialized after
every pass, which meant that it would never reach the maximum step size.
Second, wrapping the lattice in a TVar at the level of 'phase' fixes the
bug that doing phase processing in only one thread didn't actually do
anything.