Fix non-threaded 'phase' implementation.
The implementation had seriously broken behaviour in that it didn't
respond to the spreadFactor AT ALL! The only thing which had an impact
on the number of nodes was the number of input points. The new 'phase'
implementation mirrors the way the threaded implementation does things.
Only the thread management and invariant guarantees have been left out.
Thus both functions have the same behaviour and the way they are doing
it seems correct.